google-universal-image-embedder/requirements.txt
absl-py==1.2.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.9.3
async-timeout==4.0.2
attrs==22.1.0
black==22.8.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
filelock==3.8.0
frozenlist==1.3.1
fsspec==2022.8.2
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
huggingface-hub==0.9.1
hydra-core==1.2.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
kaggle==1.5.12
Markdown==3.4.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
numpy==1.23.2
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pytorch-lightning==1.7.4
pytube==12.1.0
PyYAML==6.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
text-unidecode==1.3
tinydb==4.7.0
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tqdm==4.64.1
transformers==4.21.2
typing_extensions==4.3.0
typer==0.6.1
urllib3==1.26.12
Werkzeug==2.2.2
yarl==1.8.1
zipp==3.8.1